aider: Converted project to SolidJS, added dependencies and updated files for SolidJS usage.
# Aider chat conversation:

USER: this code currently takes in a birth day, time, and location and calculates astrological data with it. there is only public/index.html, package.json, and server.js. {
  "name": "frust",
  "version": "1.0.0",
  "description": "",
  "main": "server.js",
  "type": "module",
  "scripts": {
    "test": "echo \"Error: no test specified\" && exit 1",
    "start": "node server.js"
  },
  "keywords": [],
  "author": "",
  "license": "ISC",
  "dependencies": {
    "circular-natal-horoscope-js": "^1.1.0",
    "cors": "^2.8.5",
    "ephemeris": "^2.0.0",
    "express": "^4.18.2"
  }
}
is the entire package.json. So this is barebones. let's turn this project into solidjs

ASSISTANT:
^C KeyboardInterrupt

USER: this code currently takes in a birth day, time, and location and calculates astrological data with it. there is only public/index.html, package.json, and server.js. {
  "name": "frust",
  "version": "1.0.0",
  "description": "",
  "main": "server.js",
  "type": "module",
  "scripts": {
    "test": "echo \"Error: no test specified\" && exit 1",
    "start": "node server.js"
  },
  "keywords": [],
  "author": "",
  "license": "ISC",
  "dependencies": {
    "circular-natal-horoscope-js": "^1.1.0",
    "cors": "^2.8.5",
    "ephemeris": "^2.0.0",
    "express": "^4.18.2"
  }
}
is the entire package.json. So this is barebones. let's turn this project into solidjs

ASSISTANT:
^C KeyboardInterrupt

USER: this code currently takes in a birth day, time, and location and calculates astrological data with it. there is only public/index.html, package.json, and server.js. {
  "name": "frust",
  "version": "1.0.0",
  "description": "",
  "main": "server.js",
  "type": "module",
  "scripts": {
    "test": "echo \"Error: no test specified\" && exit 1",
    "start": "node server.js"
  },
  "keywords": [],
  "author": "",
  "license": "ISC",
  "dependencies": {
    "circular-natal-horoscope-js": "^1.1.0",
    "cors": "^2.8.5",
    "ephemeris": "^2.0.0",
    "express": "^4.18.2"
  }
}
is the entire package.json. So this is barebones. let's turn this project into solidjs

ASSISTANT:
^C KeyboardInterrupt

USER: this code currently takes in a birth day, time, and location and calculates astrological data with it. there is only public/index.html, package.json, and server.js. {
  "name": "frust",
  "version": "1.0.0",
  "description": "",
  "main": "server.js",
  "type": "module",
  "scripts": {
    "test": "echo \"Error: no test specified\" && exit 1",
    "start": "node server.js"
  },
  "keywords": [],
  "author": "",
  "license": "ISC",
  "dependencies": {
    "circular-natal-horoscope-js": "^1.1.0",
    "cors": "^2.8.5",
    "ephemeris": "^2.0.0",
    "express": "^4.18.2"
  }
}
is the entire package.json. So this is barebones. let's turn this project into solidjs

ASSISTANT: To convert this project into a SolidJS project, we need to install SolidJS and its related packages, and then refactor the HTML and JavaScript code to use SolidJS syntax and conventions.

Here are the steps:

1. Install SolidJS and its related packages. We will add these dependencies to the `package.json` file.

2. Refactor the `public/index.html` file to use SolidJS syntax and conventions.

3. Refactor the `server.js` file to serve the SolidJS app.

Let's start with the first step.

```json
package.json
<<<<<<< ORIGINAL
  "dependencies": {
    "circular-natal-horoscope-js": "^1.1.0",
    "cors": "^2.8.5",
    "ephemeris": "^2.0.0",
    "express": "^4.18.2"
  }
=======
  "dependencies": {
    "circular-natal-horoscope-js": "^1.1.0",
    "cors": "^2.8.5",
    "ephemeris": "^2.0.0",
    "express": "^4.18.2",
    "solid-js": "^1.1.3",
    "vite": "^2.6.4",
    "vite-plugin-solid": "^1.2.0"
  }
>>>>>>> UPDATED
```

After updating the `package.json` file, run `npm install` to install the new dependencies.

Next, we will refactor the `public/index.html` file to use SolidJS syntax and conventions. This will involve creating a new `src/App.jsx` file and updating the `public/index.html` file to load the SolidJS app.

I will provide the changes for these files in the next steps.